Support running Dgraph Bulk in Docker (#130)
* Move Dgraph process wrapper into core prairiedog codebase

* Move "snakemake" back to requirements.txt now that datrie has been updated on pypi

* Fix conftest change

* New CircleCI workflow to also build Docker images

* New CircleCI workflow to also build Docker images

* Fix CircleCI workflows

* Fix CircleCI workflows

* Cleanup workflow, swap tag for sha

* Fix conftext.py by creating a _dg()

* Upgrade pip before installing deps

* Add a dgraph command into Snakemake

* Update Dockerfile to do lemongraph installs

* Add "delete" and "output_folder" params to DgraphBundled

* Move Dgraph process wrapper into core prairiedog codebase

* Move "snakemake" back to requirements.txt now that datrie has been updated on pypi

* Fix conftest change

* New CircleCI workflow to also build Docker images

* New CircleCI workflow to also build Docker images

* Fix CircleCI workflows

* Fix CircleCI workflows

* Cleanup workflow, swap tag for sha

* Fix conftext.py by creating a _dg()

* Upgrade pip before installing deps

* Add a dgraph command into Snakemake

* Update Dockerfile to do lemongraph installs

* Add "delete" and "output_folder" params to DgraphBundled

* Fix lemongraph install in Dockerfile

* Entrypoint into Snakemake looks good, add a test step against built docker image

* Add a .dockerignore

* Updated .dockerignore

* Fix mic df in GraphRef, and add docker run command to readme.rst

* Note "dgraph" rule in readme run cmd

* Cant ignore .git because we need to init the submodules

* Fix abs path for host volumes

* Fix perms error when calling mkdir

* Set WORKDIR for Dockerfile to /p/

* Set WORKDIR for Dockerfile to /p/

* Include custom go bins

* Switch to https for gobin submodule

* Change dgraph subprocesses to pipe depending on log level, and copy test fastas for docker image test

* Change dgraph subprocesses to pipe depending on log level, and copy test fastas for docker image test

* Revert "Change dgraph subprocesses to pipe depending on log level, and copy test fastas for docker image test"

This reverts commit 2201d71b5a9052a6efdf64bfca1b5f9d27f3468b.

* Revert "Change dgraph subprocesses to pipe depending on log level, and copy test fastas for docker image test"

This reverts commit 62d49d793377d057b10b03e2debe7b0b152902d2.

* Only pipe stderr

* Define ENTRYPOINT to use all cores

* Rm datrie install from docs

+++ prairiedog/dgraph_bundled.py
@@ -3,6 +3,7 @@ import tempfile
 import shutil
 import logging
 import time
+import pathlib
 
 import pydgraph
 
@@ -17,7 +18,7 @@ with open("dgraph/kmers.schema") as f:
     KMERS_SCHEMA = ''.join(line for line in f)
 
 
-class DG(Dgraph):
+class DgraphBundled(Dgraph):
     """
     Helper to setup and tear-down dgraph
     """
@@ -31,7 +32,7 @@ class DG(Dgraph):
         self._p_zero = subprocess.Popen(
             ['dgraph', 'zero', '-o', str(offset)], cwd=self.tmp_dir,
             stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL
+            stderr=subprocess.PIPE
         )
         time.sleep(2)
         self._p_alpha = subprocess.Popen(
@@ -39,12 +40,12 @@ class DG(Dgraph):
              'localhost:{}'.format(port("ZERO", offset)),
              '-o', str(offset)], cwd=self.tmp_dir,
             stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL
+            stderr=subprocess.PIPE
         )
         time.sleep(4)
 
     def set_schema(self):
-        self.client.alter(pydgraph.Operation(schema=DG.SCHEMA))
+        self.client.alter(pydgraph.Operation(schema=DgraphBundled.SCHEMA))
         self.client.alter(pydgraph.Operation(schema=KMERS_SCHEMA))
 
     def shutdown_dgraph(self):
@@ -53,8 +54,13 @@ class DG(Dgraph):
         self._p_zero.terminate()
         time.sleep(2)
 
-    def __init__(self):
-        self.tmp_dir = tempfile.mkdtemp()
+    def __init__(self, delete: bool = True, output_folder: str = None):
+        self.delete = delete
+        if output_folder is None:
+            self.tmp_dir = tempfile.mkdtemp()
+        else:
+            self.tmp_dir = pathlib.Path(output_folder)
+            self.tmp_dir.mkdir(parents=True, exist_ok=True)
         log.info("Will setup Dgraph from folder {}".format(self.tmp_dir))
         self._p_zero = None
         self._p_alpha = None
@@ -66,8 +72,10 @@ class DG(Dgraph):
         self.set_schema()
 
     def __del__(self):
-        self.clear()
+        if self.delete:
+            self.clear()
         time.sleep(2)
         self.shutdown_dgraph()
-        shutil.rmtree(self.tmp_dir)
+        if self.delete:
+            shutil.rmtree(self.tmp_dir)
         super().__del__()
